Use the short month name in the team calendar's away count
The full name is truncated away in the narrow column under each person's name -
"3 days off in Septem..." - and the period label above the calendar already
carries the month and the year in full.

+++ hrx2026/webapp/controller/TeamCalendar.controller.js
@@ -228,9 +228,10 @@ sap.ui.define([
 
 			oViewModel.setProperty("/busy", true);
 			oViewModel.setProperty("/periodLabel", formatter.dateRange(oFirstDay, oLastDay));
-			// The month alone: the period label above the calendar already carries the
-			// year, and this one sits in a narrow column under each name.
-			oViewModel.setProperty("/periodMonth", oFirstDay.toLocaleDateString("en-GB", { month: "long" }));
+			// Short, and without the year: this goes in the narrow column under each
+			// name, where the full month name is truncated away, and the period label
+			// above the calendar already carries the year.
+			oViewModel.setProperty("/periodMonth", oFirstDay.toLocaleDateString("en-GB", { month: "short" }));
 
 			return this._getJson(TEAM_SERVICE + "?cmd=team&" + new URLSearchParams({
 				OrgID: this._sOrgId,
